Fixing file delete bug

+++ src/helpers/deleteFile.ts
@@ -2,7 +2,9 @@ import fs from "fs";
 
 export function deleteFile(filepath: string) {
 	try {
-		fs.unlinkSync(filepath);
+		if(fs.existsSync(filepath)) {
+			fs.unlinkSync(filepath);
+		}
 	} catch(error) {
 		throw new Error(error);
 	}
